Blink用の動作確認をしやすくした。
1.Blinkの動作確認を3つの方法で出来るようにした。
* インストールされているブラウザのデータを直接参照する。
* テストデータを生成して参照させる。
* 過去に収集したCookiesファイルを所定の場所(UnitTests/Datas/*.blink)に置いて参照させる。

テストデータ生成時に暗号データの扱う関係でUtilityに復号化だけでなく暗号化機能も追加した。

+++ Net4.5/SnkLib.App.CookieGetter/Getters/GeckoCookieGetter.cs
@@ -49,7 +49,7 @@ namespace SunokoLibrary.Application.Browsers
                 Name = (string)data[1],
                 Domain = (string)data[2],
                 Path = (string)data[3],
-                Expires = Utility.UnixTimeToDateTime((int)(long)data[4]),
+                Expires = Utility.UnixTimeToDateTime((ulong)(long)data[4]),
             };
             return cookie;
         }
